+++ src/Services/MarketDataService/DataTransferObjects/StockCandleStickData.php
@@ -10,13 +10,13 @@ class StockCandleStickData
 
     public function __construct(
         int $startTimestamp,
-        public 	float $open,
-        public 	float $high,
-        public 	float $low,
-        public 	float $close,
+        public 	int $open,
+        public 	int $high,
+        public 	int $low,
+        public 	int $close,
         public 	?int $volume,
         public 	?int $numberOfTransactions,
-        public 	?float $volumeWeightedAveragePrice,
+        public 	?int $volumeWeightedAveragePrice,
     ) {
         $this->startTime = Carbon::createFromTimestampMs($startTimestamp);
     }
